+++ simulation/halsim_ws_client/src/main/native/cpp/HALSimWSClientConnection.cpp
@@ -10,22 +10,20 @@
 #include <wpi/raw_ostream.h>
 #include <wpi/raw_uv_ostream.h>
 
-#include "HALSimWSClient.h"
+#include "HALSimWS.h"
 
 namespace uv = wpi::uv;
 
-namespace wpilibws {
+using namespace wpilibws;
 
 void HALSimWSClientConnection::Initialize() {
   // Get a shared pointer to ourselves
   auto self = this->shared_from_this();
 
-  auto hws = HALSimWS::GetInstance();
-  std::string reqHost =
-      hws->GetTargetHost() + ":" + std::to_string(hws->GetTargetPort());
-
   auto ws =
-      wpi::WebSocket::CreateClient(*m_stream, hws->GetTargetUri(), reqHost);
+      wpi::WebSocket::CreateClient(*m_stream, m_client->GetTargetUri(),
+                                   wpi::Twine{m_client->GetTargetHost()} + ":" +
+                                       wpi::Twine{m_client->GetTargetPort()});
 
   ws->SetData(self);
 
@@ -35,21 +33,7 @@ void HALSimWSClientConnection::Initialize() {
   m_websocket->open.connect_extended([this](auto conn, wpi::StringRef) {
     conn.disconnect();
 
-    m_buffers = std::make_unique<BufferPool>();
-
-    m_exec =
-        UvExecFunc::Create(m_stream->GetLoop(), [](auto out, LoopFunc func) {
-          func();
-          out.set_value();
-        });
-
-    auto hws = HALSimWS::GetInstance();
-    if (!hws) {
-      wpi::errs() << "Unable to get hws instance\n";
-      return;
-    }
-
-    if (!hws->RegisterWebsocket(shared_from_this())) {
+    if (!m_client->RegisterWebsocket(shared_from_this())) {
       wpi::errs() << "Unable to register websocket\n";
       return;
     }
@@ -59,8 +43,7 @@ void HALSimWSClientConnection::Initialize() {
   });
 
   m_websocket->text.connect([this](wpi::StringRef msg, bool) {
-    auto hws = HALSimWS::GetInstance();
-    if (!m_ws_connected || !hws) {
+    if (!m_ws_connected) {
       return;
     }
 
@@ -75,7 +58,7 @@ void HALSimWSClientConnection::Initialize() {
       return;
     }
 
-    hws->OnNetValueChanged(j);
+    m_client->OnNetValueChanged(j);
   });
 
   m_websocket->closed.connect([this](uint16_t, wpi::StringRef) {
@@ -83,10 +66,7 @@ void HALSimWSClientConnection::Initialize() {
       wpi::outs() << "HALSimWS: Websocket Disconnected\n";
       m_ws_connected = false;
 
-      auto hws = HALSimWS::GetInstance();
-      if (hws) {
-        hws->CloseWebsocket(shared_from_this());
-      }
+      m_client->CloseWebsocket(shared_from_this());
     }
   });
 }
@@ -98,25 +78,24 @@ void HALSimWSClientConnection::OnSimValueChanged(const wpi::json& msg) {
   wpi::SmallVector<uv::Buffer, 4> sendBufs;
   wpi::raw_uv_ostream os{sendBufs, [this]() -> uv::Buffer {
                            std::lock_guard lock(m_buffers_mutex);
-                           return m_buffers->Allocate();
+                           return m_buffers.Allocate();
                          }};
 
   os << msg;
 
   // Call the websocket send function on the uv loop
-  m_exec->Call([this, sendBufs]() mutable {
-    m_websocket->SendText(sendBufs, [this](auto bufs, wpi::uv::Error err) {
-      {
-        std::lock_guard lock(m_buffers_mutex);
-        m_buffers->Release(bufs);
-      }
+  m_client->GetExec().Send([self = shared_from_this(), sendBufs] {
+    self->m_websocket->SendText(sendBufs,
+                                [self](auto bufs, wpi::uv::Error err) {
+                                  {
+                                    std::lock_guard lock(self->m_buffers_mutex);
+                                    self->m_buffers.Release(bufs);
+                                  }
 
-      if (err) {
-        wpi::errs() << err.str() << "\n";
-        wpi::errs().flush();
-      }
-    });
+                                  if (err) {
+                                    wpi::errs() << err.str() << "\n";
+                                    wpi::errs().flush();
+                                  }
+                                });
   });
 }
-
-}  // namespace wpilibws
