+++ simulation/halsim_ws_client/src/main/native/include/HALSimWSClientConnection.h
@@ -8,14 +8,19 @@
 #pragma once
 
 #include <memory>
+#include <utility>
 
 #include <HALSimBaseWebSocketConnection.h>
 #include <wpi/WebSocket.h>
-#include <wpi/json.h>
 #include <wpi/mutex.h>
-#include <wpi/uv/AsyncFunction.h>
 #include <wpi/uv/Buffer.h>
 #include <wpi/uv/Stream.h>
+
+#include "HALSimWS.h"
+
+namespace wpi {
+class json;
+}  // namespace wpi
 
 namespace wpilibws {
 
@@ -25,25 +30,24 @@ class HALSimWSClientConnection
     : public HALSimBaseWebSocketConnection,
       public std::enable_shared_from_this<HALSimWSClientConnection> {
  public:
-  using BufferPool = wpi::uv::SimpleBufferPool<4>;
-  using LoopFunc = std::function<void(void)>;
-  using UvExecFunc = wpi::uv::AsyncFunction<void(LoopFunc)>;
-
-  explicit HALSimWSClientConnection(std::shared_ptr<wpi::uv::Stream> stream)
-      : m_stream(stream) {}
+  explicit HALSimWSClientConnection(std::shared_ptr<HALSimWS> client,
+                                    std::shared_ptr<wpi::uv::Stream> stream)
+      : m_client(std::move(client)),
+        m_stream(std::move(stream)),
+        m_buffers(128) {}
 
  public:
   void OnSimValueChanged(const wpi::json& msg) override;
   void Initialize();
 
  private:
+  std::shared_ptr<HALSimWS> m_client;
   std::shared_ptr<wpi::uv::Stream> m_stream;
 
   bool m_ws_connected = false;
   wpi::WebSocket* m_websocket = nullptr;
 
-  std::shared_ptr<UvExecFunc> m_exec;
-  std::unique_ptr<BufferPool> m_buffers;
+  wpi::uv::SimpleBufferPool<4> m_buffers;
   std::mutex m_buffers_mutex;
 };
